Add build.sh helper script
I want links without the `.html` extension in them so I'm omitting that
part of the filename except for `index.html` and `404.html`. Because I'm
omitting `.html` I'm also adding the `_includes.head.html` to provide
one additional meta tag regarding `Content-Type`.. otherwise it's a
direct copy of `minima/_includes/head.html`.

Also ignore `tags` as it's from my Vim setup generating tags and is not
part of this project and even less so the deployment.