Add post about sanity checking post-`git rebase`
This post explains how you can be confident you didn't mess any of the
conflicts up when rebasing.

Since this is my 2nd post I had to do a minor styling change for the
list of posts shown on the startpage, also added styling to ordered
lists to become more readable.